+++ chroniclepy/src/summarising.py
@@ -2,13 +2,19 @@ from chroniclepy import utils, summarise_person
 from datetime import datetime, timedelta
 from collections import Counter
 from pytz import timezone
+import dateutil.parser
 import pandas as pd
 import numpy as np
 import os
 import re
 
-def summary(infolder, outfolder, includestartend=False, recodefile=None, fullapplistfile=None, quarterly = False, splitweek = True, weekdefinition = 'weekdayMF'):
-
+def summary(infolder, outfolder, includestartend=False, recodefile=None, 
+    fullapplistfile=None, quarterly = False, 
+    splitweek = True, weekdefinition = 'weekdayMF',
+    splitday = False, daytime = "10:00", nighttime = "22:00",
+    maxdays = None
+    ):
+        
     if not os.path.exists(outfolder):
         os.mkdir(outfolder)
 
@@ -24,7 +30,7 @@ def summary(infolder, outfolder, includestartend=False, recodefile=None, fullapp
             date_parser = lambda x: pd.to_datetime(x.rpartition('-')[0]),
             ).dropna(subset=['app_fullname'])
         allapps = allapps.union(set(preprocessed['app_fullname']))
-        personID = "-".join(str(filenm).split(".")[-2].split("-")[1:]).replace("_preprocessed","")
+        personID = str(filenm).replace("ChronicleData_preprocessed_","").replace(".csv", "")
         person = summarise_person.summarise_person(
             preprocessed,
             personID = personID,
@@ -32,7 +38,11 @@ def summary(infolder, outfolder, includestartend=False, recodefile=None, fullapp
             splitweek = splitweek,
             weekdefinition = weekdefinition,
             recodefile = recodefile,
-            includestartend = includestartend
+            includestartend = includestartend,
+            splitday = splitday,
+            daytime = daytime,
+            nighttime = nighttime,
+            maxdays = maxdays
             )
         for k,v in person.items():
             if not k in full.keys():
@@ -41,15 +51,19 @@ def summary(infolder, outfolder, includestartend=False, recodefile=None, fullapp
                 full[k] = pd.concat([full[k],person[k]],sort=True)
 
     aggfuncs = {
-        "daily":            ['mean','std'],
-        "week":             ['mean','std'],
-        "weekend":          ['mean','std'],
-        "appcoding_daily":  ['mean','std'],
-        'quarterly':        ['mean','std'],
-        "hourly":           ['mean'],
-        'appcoding_hourly': ['mean'],
-        'appcoding_week':   ['mean'],
-        'appcoding_weekend':['mean']
+        "daily":                ['mean','std'],
+        "week":                 ['mean','std'],
+        "weekend":              ['mean','std'],
+        "appcoding_daily":      ['mean','std'],
+        'quarterly':            ['mean','std'],
+        "hourly":               ['mean'],
+        'appcoding_hourly':     ['mean'],
+        'appcoding_week':       ['mean'],
+        'appcoding_weekend':    ['mean'],
+        'daytime':              ['mean', 'std'],
+        'nighttime':            ['mean', 'std'],
+        'appcoding_daytime':    ['mean'],
+        'appcoding_nighttime':  ['mean']
     }
 
     # run over all datasets, summarise and save