+++ chroniclepy/CHANGES.md
@@ -16,3 +16,9 @@
 1.2-rc2 - 07/09/2019
 * only shut off app when another one is moved to foreground (not background)
 
+1.2-rc3 - 07/31/2019
+* fix bug based on new backend
+
+1.3 - 08/08/2019
+* add possibility to split by daytime / nighttime
+* add option for maximum number of days